Use GetAllByCap to get list of network plugins in docker info
Swarmkit has enabled the plugin filtering logic based on the drivers
returned in "docker info". Currently info endpoint isnt using the
GetAllByCap, but relies on existing networks to get the plugin names.
This causes a basic issue when it comes to global network plugins which
swarm-mode relies on, wherein swarmkit will not be able to schedule the
network on the worker nodes due to the filtering logic.
In order to break this chicken & egg issue, we must start to use the
GetAllByCap instead of relying on the presence of networks to get the
plugin names.

Signed-off-by: Madhu Venugopal <madhu@docker.com>