Use multiplexed stream format for ws attach
Before this change, there is no way to differentiate stdout/stderr on
the websocket attach endpoint. With this change, the websocket attach
endpoint uses the same stream format as the non-websocket endpoint. This
allows websocket clients to differentiate stdout and stderr streams.

This continues to serve the old format for older clients.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>